Migrated aptWX -> airport
